+++ nRF Toolbox/Profiles/HeartRateMonitor/Models/HeartRateMeasurementCharacteristic.swift
@@ -14,12 +14,22 @@ private extension Flag {
 
 struct HeartRateMeasurementCharacteristic {
     let heartRate: Int
+    let date: Date
     
-    init(with data: Data) {
+    init(with data: Data, date: Date) {
+        self.date = date
+
         let flags: UInt8 = data.read()
         
         heartRate = Flag.isAvailable(bits: flags, flag: .heartRateType)
             ? Int(data.read(fromOffset: 1) as UInt16)
             : Int(data.read(fromOffset: 1) as UInt8)
     }
+
+    #if DEBUG
+    init(value: Double) {
+        self.heartRate = Int(value)
+        self.date = Date()
+    }
+    #endif
 }
